# small improvements to (de)serialize perf. helps #22593