```
julia> x=IOBuffer();
julia> a = [Int[] for i = 1:100000];
```

0.6:
```
julia> @time (seek(x,0);serialize(x, a));
  0.041324 seconds (99.52 k allocations: 6.856 MiB)

julia> @time (seek(x,0);deserialize(x));
  0.378439 seconds (1.20 M allocations: 61.010 MiB, 9.42% gc time)
```

Master:
```
julia> @time (seek(x,0);serialize(x, a));
  0.040443 seconds (99.52 k allocations: 6.856 MiB)

julia> @time (seek(x,0);deserialize(x));
  0.190012 seconds (498.51 k allocations: 21.337 MiB, 4.74% gc time)
```

That improvement is due to #22635.

This PR:
```
julia> @time (seek(x,0);serialize(x, a));
  0.029831 seconds (99.51 k allocations: 4.020 MiB)

julia> @time (seek(x,0);deserialize(x));
  0.176677 seconds (498.50 k allocations: 18.500 MiB)
```

These are some very modest improvements that can be done in a totally non-breaking way. The next thing to do is to use a `Vector` instead of an `ObjectIdDict` for backreferences when deserializing, however that could potentially break packages if they're accessing Serializer internals directly. Hopefully nobody is doing that.